fix hypervisor casing to match CloudStack enum; merge unit test into acceptance test file
