+++ cloudstack/resource_cloudstack_host.go
@@ -50,7 +50,7 @@ func resourceCloudStackHost() *schema.Resource {
 			"hypervisor": {
 				Type:         schema.TypeString,
 				Required:     true,
-				ValidateFunc: validation.StringInSlice([]string{"xenserver", "kvm", "vmware", "baremetal", "simulator"}, false),
+				ValidateFunc: validation.StringInSlice([]string{"XenServer", "KVM", "VMware", "Hyperv", "BareMetal", "Simulator", "Ovm3"}, false),
 				ForceNew:     true,
 			},
 			"pod_id": {
